+++ src/io/graphviz.i.hpp
@@ -257,7 +257,7 @@ void graphviz_deserialize(
           inserted_vertices[source_vertex_name] = source;
         }
 
-        for (int64_t i = 1; i < edge_vertices.size(); ++i) {
+        for (size_t i = 1; i < edge_vertices.size(); ++i) {
           std::string target_vertex_name = utils::trim(edge_vertices[i]);
           if (!target_vertex_name.empty()) {
